+++ objc2/examples/delegate.rs
@@ -13,7 +13,7 @@ extern_class!(
     struct NSResponder;
 
     unsafe impl ClassType for NSResponder {
-        type Superclass = NSObject;
+        type Super = NSObject;
     }
 );
 
@@ -26,13 +26,13 @@ declare_class!(
 
     unsafe impl ClassType for CustomAppDelegate {
         #[inherits(NSObject)]
-        type Superclass = NSResponder;
+        type Super = NSResponder;
     }
 
     unsafe impl CustomAppDelegate {
         #[sel(initWith:another:)]
         fn init_with(self: &mut Self, ivar: u8, another_ivar: Bool) -> *mut Self {
-            let this: *mut Self = unsafe { msg_send![super(self, NSResponder::class()), init] };
+            let this: *mut Self = unsafe { msg_send![super(self), init] };
             if let Some(this) = unsafe { this.as_mut() } {
                 // TODO: Allow initialization through MaybeUninit
                 *this.ivar = ivar;