+++ test-ui/ui/msg_send_super_not_classtype.rs
@@ -1,0 +1,12 @@
+//! Invalid receiver to msg_send![super(obj), ...], missing ClassType impl.
+use objc2::msg_send;
+use objc2::foundation::NSObject;
+use objc2::runtime::Object;
+
+fn main() {
+    let obj: &Object;
+    let _: () = unsafe { msg_send![super(obj), method] };
+
+    let obj: &NSObject; // impls ClassType, but it's superclass does not
+    let _: () = unsafe { msg_send![super(obj), method] };
+}