+++ objc2/CHANGELOG.md
@@ -19,8 +19,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   methods that `extern_class!` and `declare_class!` generated to that. This
   means you'll have to `use objc2::ClassType` whenever you want to use e.g.
   `NSData::class()`.
-* Added `Id::into_superclass`.
+* Added `Id::into_super`.
 * Added `extern_methods!` macro.
+* Added ability to call `msg_send![super(obj), ...]` without explicitly
+  specifying the superclass.
 
 ### Changed
 * **BREAKING**: Change syntax in `extern_class!` macro to be more Rust-like.